# src/main/java/com/sairaj/expense/tracker/dto/TokenResponse.java
package com.sairaj.expense.tracker.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

@Data @Builder @AllArgsConstructor
public class TokenResponse {
    String token;
}