# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  version: 1.0.0

servers:
  - url: http://localhost:9090
    description: Local server

tags:
  - name: Auth
    description: APIs used for authentication and user management
  - name: Expense
    description: APIs used to manage expense management

paths:
  /api/auth/create:
    post:
      tags:
        - Auth
      summary: Create User
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User Created
          content:
            application/json:
              schema:
                type: string
                example: "User Created"
        '400':
          description: Bad Request, wrong request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict, user already exists in system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login User
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request, wrong request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized, wrong credentials used

components:
  schemas:
    TokenResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: saul@goodman.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 64
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z\d]).{8,}$'
          example: Password@1234
    ErrorResponse:
      type: object
      required:
        - message
        - date
      properties:
        message:
          type: string
          example: Details about error
        date:
          type: string
          format: date
          example: Date at which error occured
    UserRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          example: Saul Goodman
        email:
          type: string
          format: email
          example: saul@goodman.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 64
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z\d]).{8,}$'
          example: Password@1234